requirements.txt
biopython
pandas
seaborn
numpy 
matplotlib